add option to disable ssl context refresh
With the way reload() on SslContextFactory in Jetty works there are
certain things cleared and never set gain (_certHost and _certWilds).
These two properties are onlyt set when load() is called for the first
time with a null context. If these properties are not correctly set then
the SNI matching logic that allows multiple certificates in your
keystore to be used and selected based off SNI does not work.